Add CMBH collector and apply changes to make it possible
Along with the new collector, this change makes a number of model fields
optional, given the data to fill them out is not available in the CMBH
web site. It also removes the hack to make the graph in the per nature
totals view be the same height as the table, since the table is now much
taller.