Major rework of narrowing down the data that is showed
Instead of allowing selection of a single legislative house, we now have
a dropdown in the top menu that allows excluding houses or legislatures.
Legislature is a new model/concept, that is like the 'class' for a mandate,
it is tied to the institution instead of being tied to the legislator.

This change also adapts the collectors to use this new structure.