Go back to showing a single institution or all
We haven't managed to figure out a good way of showing multiple institutions in
an easy to grasp way, and this is causing more grief and confusion than being helpful,
so go back to showing a single institution at a time (while keeping 'all' viable).

This change migrates the per nature view to the old new model.